Split pretty hostname functions into their own file.
